pyproject.toml
[tool.poetry]
name = "pubq"
version = "1.0.0"
description = "PUBQ Python SDK"
license = "MIT"
authors = ["PUBQ Team <dev@pubq.io>"]
maintainers = ["PUBQ Team <dev@pubq.io>"]
readme = "README.md"
homepage = "https://pubq.io"
repository = "https://github.com/pubqio/pubq-python"
documentation = "https://github.com/pubqio/pubq-python"
classifiers = [
    "Development Status :: 6 - Mature",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.8"
pyee = "^11.0.0"
socketclusterclient = "^1.3.6"
requests = "^2.31.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
